% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Edwards_2015}
\alias{Edwards_2015}
\title{This is the Edwards dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 32 rows and 20 columns.
}
\usage{
data(Edwards_2015)
}
\description{
This is the Edwards dataset
}
\author{
Edwards et al.
}
\keyword{datasets}
