% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Edwards_cleaned}
\alias{Edwards_cleaned}
\title{This is the Edwards cleaned dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 256 rows and 8 columns.
}
\usage{
data(Edwards_cleaned)
}
\description{
This is the Edwards cleaned dataset
}
\author{
Edwards et al.
}
\keyword{datasets}
