% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gysin_cleaned}
\alias{Gysin_cleaned}
\title{This is the Gysin cleaned dataset}
\format{
An object of class \code{data.frame} with 120 rows and 13 columns.
}
\source{
\url{https://datadryad.org/stash/dataset/doi:10.5061/dryad.85nf3}
}
\usage{
data(Gysin_cleaned)
}
\description{
Gysin-Maillart, A., Schwab, S., Soravia, L., Megert, M., & Michel, K. (2016).
A Novel Brief Therapy for Patients Who Attempt Suicide: A 24-months Follow-Up
Randomized Controlled Study of the Attempted Suicide Short Intervention Program (ASSIP).
PLOS Medicine, 13(3), e1001968. https://doi.org/10.1371/journal.pmed.1001968
}
\author{
Gysin et al.
}
\keyword{datasets}
