% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Larson_2015}
\alias{Larson_2015}
\title{This is the Larson dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 135 rows and 15 columns.
}
\usage{
data(Larson_2015)
}
\description{
This is the Larson dataset
}
\author{
Larson et al.
}
\keyword{datasets}
