% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{McErlean_2020}
\alias{McErlean_2020}
\title{This is the McErlean_2020 dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 179 rows and 129 columns.
}
\usage{
data(McErlean_2020)
}
\description{
This is the McErlean_2020 dataset
}
\author{
McErlean et al.
}
\keyword{datasets}
