% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{McErlean_cleaned}
\alias{McErlean_cleaned}
\title{This is the McErlean_cleaned dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 177 rows and 17 columns.
}
\usage{
data(McErlean_cleaned)
}
\description{
This is the McErlean_cleaned dataset
}
\author{
McErlean et al.
}
\keyword{datasets}
