% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Palser_2018}
\alias{Palser_2018}
\title{This is the Palser_2018 dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 384 rows and 9 columns.
}
\usage{
data(Palser_2018)
}
\description{
This is the Palser_2018 dataset
}
\author{
Palser et al.
}
\keyword{datasets}
