% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Palser_cleaned}
\alias{Palser_cleaned}
\title{This is the Palser_cleaned dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 384 rows and 29 columns.
}
\usage{
data(Palser_cleaned)
}
\description{
This is the Palser_cleaned dataset
}
\author{
Palser et al.
}
\keyword{datasets}
