% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Palser_tas_20}
\alias{Palser_tas_20}
\title{This is the Palser_tas_20 dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 384 rows and 21 columns.
}
\usage{
data(Palser_tas_20)
}
\description{
This is the Palser_tas_20 dataset
}
\author{
Palser et al.
}
\keyword{datasets}
