% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Reilly_2016}
\alias{Reilly_2016}
\title{This is the Reilly_2016 dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 313 rows and 27 columns.
}
\usage{
data(Reilly_2016)
}
\description{
This is the Reilly_2016 dataset
}
\author{
Reilly et al.
}
\keyword{datasets}
