% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tas_as_text}
\alias{tas_as_text}
\title{This is the tas_as_text dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 242 rows and 22 columns.
}
\usage{
data(tas_as_text)
}
\description{
This is the tas_as_text dataset
}
\author{
Rice & Ryan
}
\keyword{datasets}
