% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vdv_2015}
\alias{vdv_2015}
\title{This is the vdv dataset}
\format{
An object of class \code{data.frame} with 293 rows and 19 columns.
}
\usage{
data(vdv_2015)
}
\description{
This is the vdv dataset
}
\author{
van der Velde et al.
}
\keyword{datasets}
