% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Branjerdporn_2019}
\alias{Branjerdporn_2019}
\title{This is the Branjerdporn_2019 dataset}
\format{
An object of class \code{data.frame} with 156 rows and 113 columns.
}
\usage{
data(Branjerdporn_2019)
}
\description{
This is the Branjerdporn_2019 dataset
}
\author{
Branjerdporn et al.
}
\keyword{datasets}
