% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Branjerdporn_2019}
\alias{Branjerdporn_2019}
\title{Branjerdporn_2019 dataset}
\format{
An object of class \code{data.frame} with 156 rows and 113 columns.
}
\source{
\url{https://ndownloader.figstatic.com/files/14059100}
}
\usage{
data(Branjerdporn_2019)
}
\description{
Branjerdporn, G., Meredith, P., Strong, J., & Green, M. (2019).
Sensory sensitivity and its relationship with adult attachment and parenting styles.
PLOS ONE, 14(1), e0209555. https://doi.org/10.1371/journal.pone.0209555.
}
\author{
Branjerdporn et al.
}
\keyword{datasets}
