% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gysin_cleaned}
\alias{Gysin_cleaned}
\title{This is the Gysin cleaned dataset}
\format{
An object of class \code{data.frame} with 120 rows and 13 columns.
}
\usage{
data(Gysin_cleaned)
}
\description{
This is the Gysin cleaned dataset
}
\author{
Gysin et al.
}
\keyword{datasets}
